Updated nf-core modules and subworkflows
