+++ modules/nf-core/custom/dumpsoftwareversions/meta.yml
