+++ modules/nf-core/fastqc/meta.yml
