--- modules/nf-core/modules/bwa/mem/meta.yml
@@ -1,50 +1,0 @@
-name: bwa_mem
-description: Performs fastq alignment to a fasta reference using BWA
-keywords:
-  - mem
-  - bwa
-  - alignment
-  - map
-  - fastq
-  - bam
-  - sam
-tools:
-  - bwa:
-      description: |
-        BWA is a software package for mapping DNA sequences against
-        a large reference genome, such as the human genome.
-      homepage: http://bio-bwa.sourceforge.net/
-      documentation: http://www.htslib.org/doc/samtools.html
-      arxiv: arXiv:1303.3997
-      licence: ["GPL-3.0-or-later"]
-input:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - reads:
-      type: file
-      description: |
-        List of input FastQ files of size 1 and 2 for single-end and paired-end data,
-        respectively.
-  - index:
-      type: file
-      description: BWA genome index files
-      pattern: "Directory containing BWA index *.{amb,ann,bwt,pac,sa}"
-  - sort_bam:
-      type: boolean
-      description: use samtools sort (true) or samtools view (false)
-      pattern: "true or false"
-output:
-  - bam:
-      type: file
-      description: Output BAM file containing read alignments
-      pattern: "*.{bam}"
-  - versions:
-      type: file
-      description: File containing software versions
-      pattern: "versions.yml"
-authors:
-  - "@drpatelh"
-  - "@jeremy1805"
