--- modules/nf-core/modules/picard/markduplicates/meta.yml
@@ -1,52 +1,0 @@
-name: picard_markduplicates
-description: Locate and tag duplicate reads in a BAM file
-keywords:
-  - markduplicates
-  - pcr
-  - duplicates
-  - bam
-  - sam
-  - cram
-tools:
-  - picard:
-      description: |
-        A set of command line tools (in Java) for manipulating high-throughput sequencing (HTS)
-        data and formats such as SAM/BAM/CRAM and VCF.
-      homepage: https://broadinstitute.github.io/picard/
-      documentation: https://broadinstitute.github.io/picard/
-      licence: ["MIT"]
-input:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - bam:
-      type: file
-      description: BAM file
-      pattern: "*.{bam}"
-output:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - bam:
-      type: file
-      description: BAM file with duplicate reads marked/removed
-      pattern: "*.{bam}"
-  - bai:
-      type: file
-      description: An optional BAM index file. If desired, --CREATE_INDEX must be passed as a flag
-      pattern: "*.{bai}"
-  - metrics:
-      type: file
-      description: Duplicate metrics file generated by picard
-      pattern: "*.{metrics.txt}"
-  - versions:
-      type: file
-      description: File containing software versions
-      pattern: "versions.yml"
-authors:
-  - "@drpatelh"
-  - "@projectoriented"
