--- modules/nf-core/modules/samtools/view/meta.yml
@@ -1,57 +1,0 @@
-name: samtools_view
-description: filter/convert SAM/BAM/CRAM file
-keywords:
-  - view
-  - bam
-  - sam
-  - cram
-tools:
-  - samtools:
-      description: |
-        SAMtools is a set of utilities for interacting with and post-processing
-        short DNA sequence read alignments in the SAM, BAM and CRAM formats, written by Heng Li.
-        These files are generated as output by short read aligners like BWA.
-      homepage: http://www.htslib.org/
-      documentation: hhttp://www.htslib.org/doc/samtools.html
-      doi: 10.1093/bioinformatics/btp352
-      licence: ["MIT"]
-input:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - input:
-      type: file
-      description: BAM/CRAM/SAM file
-      pattern: "*.{bam,cram,sam}"
-  - index:
-      type: optional file
-      description: BAM.BAI/CRAM.CRAI file
-      pattern: "*.{.bai,.crai}"
-  - fasta:
-      type: optional file
-      description: Reference file the CRAM was created with
-      pattern: "*.{fasta,fa}"
-output:
-  - meta:
-      type: map
-      description: |
-        Groovy Map containing sample information
-        e.g. [ id:'test', single_end:false ]
-  - bam:
-      type: file
-      description: filtered/converted BAM/SAM file
-      pattern: "*.{bam,sam}"
-  - cram:
-      type: file
-      description: filtered/converted CRAM file
-      pattern: "*.cram"
-  - versions:
-      type: file
-      description: File containing software versions
-      pattern: "versions.yml"
-authors:
-  - "@drpatelh"
-  - "@joseespinosa"
-  - "@FriederikeHanssen"
