+++ subworkflows/nf-core/bam_markduplicates_picard/meta.yml
@@ -1,0 +1,61 @@
+name: "bam_markduplicates_picard"
+description: Picard MarkDuplicates, index BAM file and run samtools stats, flagstat and idxstats
+keywords:
+  - markduplicates
+  - bam
+  - sam
+  - cram
+
+modules:
+  - picard/markduplicates
+  - samtools/index
+  - samtools/stats
+  - samtools/idxstats
+  - samtools/flagstat
+
+input:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test' ]
+  - bam:
+      type: file
+      description: BAM/CRAM/SAM file
+      pattern: "*.{bam,cram,sam}"
+  - fasta:
+      type: file
+      description: Reference genome fasta file
+      pattern: "*.{fasta,fa}"
+
+output:
+  - meta:
+      type: map
+      description: |
+        Groovy Map containing sample information
+        e.g. [ id:'test' ]
+  - bai:
+      type: file
+      description: BAM/CRAM/SAM samtools index
+      pattern: "*.{bai,crai,sai}"
+  - csi:
+      type: file
+      description: CSI samtools index
+      pattern: "*.csi"
+  - stats:
+      type: file
+      description: File containing samtools stats output
+  - flagstat:
+      type: file
+      description: File containing samtools flagstat output
+  - idxstats:
+      type: file
+      description: File containing samtools idxstats output
+      pattern: "*.{idxstats}"
+  - versions:
+      type: file
+      description: File containing software versions
+      pattern: "versions.yml"
+authors:
+  - "@dmarron"
+  - "@drpatelh"
